+++ src/utils/progress.hpp
@@ -4,11 +4,11 @@
 
 class Progress
 {
-    unsigned int m_current;
-    unsigned int m_total;
+    size_t m_current;
+    size_t m_total;
 public:
-    Progress(unsigned int total);
+    Progress(size_t total);
     std::string to_string() const;
-    void increment(unsigned int value = 1);
-    unsigned int total() const;
+    void increment(size_t value = 1);
+    size_t total() const;
 };
